split gravitational dynamics and gravity field interfaces
